BUG: Fix a problem with a specific Hitachi scanner where a volume was created for each image of each series.
This scanner creates a multiframe file of one frame for each image, thus leading to a volume for each image in Starviewer. With this change we detect if an image is from this scanner, and if that's the case we use a different logic where the volume number in multiframe and enhanced images is incremented only with the same criteria as for single frame images.

This fixes #2112.