ENH: SlicingMouseTool with polar sensitive axis.
Depending on the direction of the movement a different axis is
modified. This means one can change phases moving horizontally, do
slicing moving vertically or another thing when moving within the
oblique axis.

The user interaction of this implementation is not successful. This
commit is here just for archiving purposes.