ENH: Responsive keyboard feedback that accumulates key press events.
Keyboard events are accumulated, after 10ms wihout keyboard events the
actual scrolling is done. When a costly scrolling operation takes place,
there might be a big number of key presses in the buffer. Thanks to the
timer, all of them are packed and thus the user feedback becomes
responsive and does not hang due to the accumulation.

The scrolling algorithm has been adapted to the floating point version.