BUG: Show display shutters.
Display shutters were never visible because when the GenericVolumeDisplayUnitHandler stack was built the VolumeDisplayUnit had not added yet m_shutterImageSlice to its own stack, and the handler stack was built only one time. The solution has been to rebuild the handler stack each time that there is a change in one of the units' stacks, i.e. when a shutter is added or enabled or disabled.