ENH: Remove an unwanted border in some images.
In some images (CTs and MGs for example), we were seeing un ugly 1 pixel border
around the image at some zoom levels. It turned out that this is the background
from the vtkImageReslice used by vtkImageResliceMapper to obtain an image the
size of the viewer. This background should remain of the the display boundary of
the actor, but likely due to some precision quirks, it just slips 1 pixel in at
some zoom levels.

This background was set at voxel value 0, but it should instead have the minimum
value of the image or less (in CT the minimum is -1024). We have opted to set it
at negative infinity to ensure that it is always black (except when using an
inverse window level, in which case we have to set it at positive infinity).