ENH: Major refactoring of ApplicationVersionChecker.
Now it only has static methods. The main ones are checkAndShowReleaseNotes() (that replaces the previous checkReleaseNotes(), showIfCorrect(), and a few private ones), and showLocalReleaseNotes(). There is also isNewerVersion, that is public to make it testable.

The advantages over the previous version are:

- No need to keep an instance to call two methods at the appropriate moments to check for new versions or show release notes; a single static method call is enough to do it all.
- Simpler interface.
- Easier to understand code.

In addition, the update check is no longer done in QApplicationMainWindow's constructor. Instead, QApplicationMainWindow now has a method to do it that is called once from main. This has the advantages that the check is only performed once in the process lifetime, and that it is not done in tests.